impr: Move Elevation components to elevation mod
