+++ ryot/src/bevy_ryot/game/mod.rs
@@ -1,12 +1,9 @@
-use core::fmt;
-use std::marker::PhantomData;
-
+use self::elevation::{apply_elevation, Elevation};
 use crate::prelude::*;
 use bevy::prelude::*;
-use itertools::Itertools;
+use core::fmt;
 use serde::{Deserialize, Serialize};
-
-use self::{elevation::Elevation, map::MapTiles, sprites::SPRITE_BASE_SIZE};
+use std::marker::PhantomData;
 
 pub mod elevation;
 
@@ -112,61 +109,3 @@ impl GameObjectBundle {
 
 #[derive(Event, Clone, Debug, Deref, DerefMut)]
 pub struct LoadObjects(pub Vec<GameObjectBundle>);
-
-type ElevationFilter = (
-    With<GameObjectId>,
-    Or<(
-        Changed<GameObjectId>,
-        Changed<Visibility>,
-        Changed<TilePosition>,
-    )>,
-);
-
-fn apply_elevation<C: AppearanceAssets>(
-    appearance_asets: Res<C>,
-    q_tile: Query<(&TilePosition, &Layer), ElevationFilter>,
-    mut q_entities: Query<(&mut Elevation, &GameObjectId, Option<&Visibility>)>,
-    map_tiles: Res<MapTiles<Entity>>,
-) {
-    let appearances = appearance_asets.prepared_appearances();
-    for tile in q_tile
-        .iter()
-        .filter(|(_, layer)| matches!(layer, Layer::Bottom(_)))
-        .map(|(pos, _)| *pos)
-        .unique()
-        .filter_map(|pos| map_tiles.get(&pos))
-    {
-        tile.into_iter()
-            .filter(|(layer, _)| matches!(layer, Layer::Bottom(_)))
-            .map(|(_, entity)| entity)
-            .fold(0., |tile_elevation, entity| {
-                let Ok((mut elevation, object_id, visibility)) = q_entities.get_mut(entity) else {
-                    return tile_elevation;
-                };
-                let Some((group, id)) = object_id.as_group_and_id() else {
-                    return tile_elevation;
-                };
-
-                let elevation_delta =
-                    if visibility.cloned().unwrap_or_default() != Visibility::Hidden {
-                        appearances
-                            .get_for_group(group, id)
-                            .cloned()
-                            .and_then(|app| app.flags?.height?.elevation)
-                            .unwrap_or(0) as f32
-                            / SPRITE_BASE_SIZE.y as f32
-                    } else {
-                        0.
-                    };
-
-                elevation.elevation = match group {
-                    AppearanceGroup::Object => tile_elevation,
-                    AppearanceGroup::Outfit => tile_elevation,
-                    AppearanceGroup::Effect => 0.,
-                    AppearanceGroup::Missile => 0.,
-                };
-
-                tile_elevation + elevation_delta
-            });
-    }
-}
